Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/Create-valid-url-protocol-empty.any-expected.txt

CONSOLE MESSAGE: line 6: WebSocket connection to 'ws://localhost:49001/echo' failed: WebSocket is closed before the connection is established.

PASS Create WebSocket - wsocket.protocol should be empty before connection is established 
